Niall has implemented displaying enemy health, it was all Niall, honest.

+++ SDL_Game_Framework/Game_Time.h
@@ -10,13 +10,13 @@ public:
 	Game_Time();
 	~Game_Time();
 
-	double Seconds_Since_Start();
-	double Milliseconds_Since_Start();
-	double Seconds_Since_Last_Call();
-	double Milliseconds_Since_Last_Call();
+	float Seconds_Since_Start();
+	float Milliseconds_Since_Start();
+	float Seconds_Since_Last_Call();
+	float Milliseconds_Since_Last_Call();
 
 private:
-	double m_last_call_time;
+	float m_last_call_time;
 };
 
 #endif //_GAME_TIME_H_
